+++ pkg/tools/otelcol-config/main_windows.go
@@ -1,0 +1,6 @@
+package main
+
+// this is only a stub for Windows build support
+func setSumologicRemoteOwner(*flagValues) error {
+	return nil
+}
